pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "echo"
version = "1.0.0"
description = "Instant accent & clarity feedback for speech clips"
authors = [{name = "Echo Team", email = "echo@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "gradio>=4.0.0,<5.0.0",
    "fastapi>=0.104.0,<1.0.0",
    "uvicorn>=0.24.0,<1.0.0",
    "pydub>=0.25.1,<1.0.0",
    "soundfile>=0.12.1,<1.0.0",
    "librosa>=0.10.1,<1.0.0",
    "webrtcvad>=2.0.10,<3.0.0",
    "ctranslate2>=3.24.0,<4.0.0",
    "faster-whisper>=0.10.0,<1.0.0",
    "vosk>=0.3.45,<1.0.0",
    "g2p-en>=2.1.0,<3.0.0",
    "jiwer>=3.0.3,<4.0.0",
    "plotly>=5.17.0,<6.0.0",
    "reportlab>=4.0.7,<5.0.0",
    "pyyaml>=6.0.1,<7.0.0",
    "pydantic>=2.5.0,<3.0.0",
    "numpy>=1.24.0,<2.0.0",
    "scipy>=1.11.0,<2.0.0",
    "numba>=0.58.0,<1.0.0",
    "mediapipe>=0.10.0,<1.0.0",
    "opencv-python>=4.8.0,<5.0.0",
    "python-multipart>=0.0.6,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<8.0.0",
    "pytest-asyncio>=0.21.1,<1.0.0",
    "black>=23.9.0,<24.0.0",
    "isort>=5.12.0,<6.0.0",
    "flake8>=6.1.0,<7.0.0",
    "mypy>=1.6.0,<2.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "webui*", "cli*"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true